+++ src/analysis/a2/combining_datafilter.h
@@ -79,7 +79,7 @@ struct CombiningFilter
 };
 
 CombiningFilter make_combining_filter(CombiningFilter::Flag flags,
-                                      u16 wordCount,
+                                      u8 wordCount,
                                       const std::vector<std::string> &filterStrings = {});
 
 bool validate(CombiningFilter *cf);
